+++ regression/reference/classes/pyclassesutil.cpp
@@ -9,6 +9,7 @@
 #include "pyclassesmodule.hpp"
 
 const char *PY_Class1_capsule_name = "Class1";
+const char *PY_Class2_capsule_name = "Class2";
 const char *PY_Singleton_capsule_name = "Singleton";
 
 
@@ -54,6 +55,50 @@ int PP_Class1_from_Object(PyObject *obj, void **addr)
     *addr = self->obj;
     return 1;
     // splicer end class.Class1.utility.from_object
+}
+
+// Wrap pointer to struct/class.
+PyObject *PP_Class2_to_Object_idtor(classes::Class2 *addr, int idtor)
+{
+    // splicer begin class.Class2.utility.to_object
+    PY_Class2 *obj = PyObject_New(PY_Class2, &PY_Class2_Type);
+    if (obj == nullptr)
+        return nullptr;
+    obj->obj = addr;
+    obj->idtor = idtor;
+    return reinterpret_cast<PyObject *>(obj);
+    // splicer end class.Class2.utility.to_object
+}
+
+// converter which may be used with PyBuild.
+PyObject *PP_Class2_to_Object(classes::Class2 *addr)
+{
+    // splicer begin class.Class2.utility.to_object
+    PyObject *voidobj;
+    PyObject *args;
+    PyObject *rv;
+
+    voidobj = PyCapsule_New(addr, PY_Class2_capsule_name, nullptr);
+    args = PyTuple_New(1);
+    PyTuple_SET_ITEM(args, 0, voidobj);
+    rv = PyObject_Call((PyObject *) &PY_Class2_Type, args, nullptr);
+    Py_DECREF(args);
+    return rv;
+    // splicer end class.Class2.utility.to_object
+}
+
+// converter which may be used with PyArg_Parse.
+int PP_Class2_from_Object(PyObject *obj, void **addr)
+{
+    // splicer begin class.Class2.utility.from_object
+    if (obj->ob_type != &PY_Class2_Type) {
+        // raise exception
+        return 0;
+    }
+    PY_Class2 * self = (PY_Class2 *) obj;
+    *addr = self->obj;
+    return 1;
+    // splicer end class.Class2.utility.from_object
 }
 
 // Wrap pointer to struct/class.