+++ regression/reference/none/helpers.c
@@ -15,7 +15,7 @@ typedef struct {
 // helper ShroudLenTrim
 // Returns the length of character string src with length nsrc,
 // ignoring any trailing blanks.
-int ShroudLenTrim(const char *src, int nsrc) {
+static int ShroudLenTrim(const char *src, int nsrc) {
     int i;
 
     for (i = nsrc - 1; i >= 0; i--) {