+++ regression/reference/strings/wrapstrings.cpp
@@ -81,20 +81,6 @@ static void ShroudStrToArray(STR_SHROUD_array *array, const std::string * src, i
     array->rank = 1;
 }
 // end helper ShroudStrToArray
-
-// start helper copy_string
-// helper copy_string
-// Copy the char* or std::string in context into c_var.
-// Called by Fortran to deal with allocatable character.
-void STR_ShroudCopyStringAndFree(STR_SHROUD_array *data, char *c_var, size_t c_var_len) {
-    const char *cxx_var = data->addr.ccharp;
-    size_t n = c_var_len;
-    if (data->elem_len < n) n = data->elem_len;
-    std::strncpy(c_var, cxx_var, n);
-    STR_SHROUD_memory_destructor(&data->cxx); // delete data->cxx.addr
-}
-// end helper copy_string
-
 // splicer begin C_definitions
 // splicer end C_definitions